Moved tile to its own file
