+++ button.py
@@ -17,7 +17,7 @@ class Button():
             if pygame.mouse.get_pressed()[0] == 1 and self.clicked == False:
                 self.clicked = True
                 action = True
-            
+
             if pygame.mouse.get_pressed()[0] == 0:
                 self.clicked = False
