+++ screen.py
@@ -1,6 +1,7 @@
 import button
 import os
 import piece
+import tile
 from pickle import TRUE
 import pygame
 import sys
@@ -14,6 +15,7 @@ TILES = [[0 for i in range(COLS)] for j in range(ROWS)]
 app_folder = os.path.dirname(os.path.realpath(sys.argv[0]))
 os.chdir(app_folder)
 
+# pygame initialization and screen generation
 pygame.init()
 background_colour = (169, 169, 169)
 (width, height) = (800, 800)
@@ -23,70 +25,14 @@ screen.fill(background_colour)
 pygame.display.update()
 running = True
 
-
-class Tile:
-    def __init__(self, xPos, yPos, color):
-        self.x = xPos
-        self.y = yPos
-        self.color = color
-        self.size = 50
-        # Active is used to check for border or corner tiles
-        self.active = True
-        print("Made a tile at: " + str(self.x) + ", " + str(self.y))
-
-    def draw(self, screen):
-        gray = (169, 169, 169)
-        black = (0, 0, 0)
-        white = (255, 255, 255)
-        blue = (0, 0, 255)
-        surface = screen
-
-        # Makes pattern on board
-        if self.color == 1:
-            pygame.draw.rect(surface, white, pygame.Rect(
-                self.x*self.size, self.y*self.size, self.size, self.size))
-        else:
-            pygame.draw.rect(surface, black, pygame.Rect(
-                self.x*self.size, self.y*self.size, self.size, self.size))
-
-        # Makes gray corners and blue outline
-        if self.x == 0 or self.y == 0 or self.x == 15 or self.y == 15:
-            pygame.draw.rect(surface, blue, pygame.Rect(
-                self.x*self.size, self.y*self.size, self.size, self.size))
-            self.active = False
-        if self.x > 0 and self.x < 4 and self.y > 0 and self.y < 4:
-            pygame.draw.rect(surface, gray, pygame.Rect(
-                self.x*self.size, self.y*self.size, self.size, self.size))
-            self.active = False
-        if self.x > 11 and self.x < 15 and self.y > 11 and self.y < 15:
-            pygame.draw.rect(surface, gray, pygame.Rect(
-                self.x*self.size, self.y*self.size, self.size, self.size))
-            self.active = False
-
-        if self.x > 0 and self.x < 4 and self.y > 11 and self.y < 15:
-            pygame.draw.rect(surface, gray, pygame.Rect(
-                self.x*self.size, self.y*self.size, self.size, self.size))
-            self.active = False
-        if self.x > 11 and self.x < 15 and self.y > 0 and self.y < 4:
-            pygame.draw.rect(surface, gray, pygame.Rect(
-                self.x*self.size, self.y*self.size, self.size, self.size))
-            self.active = False
-        # For some reason this draws to the screen, don't ask why I have no idea yet
-        pygame.display.update()
-
-    def getPos(self):
-        print(self.x + ", " + self.y)
-
 # creates the chess board
-
-
 def createBoard(running):
     board_made = False
     while running:
         if board_made is False:
             for x in range(ROWS):
                 for y in range(COLS):
-                    TILES[y][x] = Tile(x, y, x % 2 + y % 2)
+                    TILES[y][x] = tile.Tile(x, y, x % 2 + y % 2)
             for x in range(ROWS):
                 for y in range(COLS):
                     TILES[y][x].draw(screen)
